more rpm NEVR/NEVRA parsing, shifted to its own module
